requirements.txt
streamlit==1.44.1
lxml==5.3.1
pandas==2.2.3
numpy==2.2.4
openai==1.68.2
faiss-cpu==1.10.0
python-dotenv==1.0.1